Move test to own class
